Move Violations pre processing behaviour into its own folder
