+++ generator/src/test/java/com/scottlogic/deg/generator/smoke_tests/ExampleProfilesViolationTests.java
@@ -3,7 +3,7 @@ package com.scottlogic.deg.generator.smoke_tests;
 import com.scottlogic.deg.generator.GenerationEngine;
 import com.scottlogic.deg.generator.Profile;
 import com.scottlogic.deg.generator.ProfileFields;
-import com.scottlogic.deg.generator.ViolationGenerationEngineWrapper;
+import com.scottlogic.deg.generator.violations.ViolationGenerationEngineWrapper;
 import com.scottlogic.deg.generator.decisiontree.MostProlificConstraintOptimiser;
 import com.scottlogic.deg.generator.decisiontree.ProfileDecisionTreeFactory;
 import com.scottlogic.deg.generator.decisiontree.tree_partitioning.RelatedFieldTreePartitioner;
